fix: resolve all code review issues across shared, desktop, cli modules
  Addresses ~50 issues identified in docs/CODE_REVIEW.md covering
  critical bugs, high-severity thread safety, medium-severity fixes,
  Kotlin idiom violations, CLI cleanups, and build file hygiene.

  Critical bugs (1.1-1.4):
  - RateLimiter: move delay() outside mutex.withLock to unblock coroutines
  - FileStatus: guard detection ratio split against IndexOutOfBounds
  - PendingRecheckTracker: replace mutableMapOf with ConcurrentHashMap
  - VTResponseParser: fix copy-paste bug where Pair had identical elements,
    replace triple-optional Pair with DetectionStats data class

  Thread safety (2.1-2.4):
  - AppContainer: add @Volatile, capture _apiKey in local val for TOCTOU fix
  - LocalTelemetry: wrap data mutations in Mutex
  - PauseController: use CopyOnWriteArrayList for callbacks
  - AppStore: use _state.update {} for atomic read-modify-write

  Medium-severity (3.1-3.42):
  - Remove dead overrides map from AppContainer
  - PendingRecheckTracker: implement Closeable, privatize callbacks
  - VirusTotalApi: implement Closeable, internal getApiKey(), remove
    redundant catch-and-rethrow
  - CredentialStore: val file, log save() failures, document Windows NTFS
  - ExtensionsConfig: synchronized file access, fall back to defaults
  - CrashLogger: bound readText to 1MB/500 lines
  - UserInfo: parameterize HttpClientEngine
  - SideEffects: .use {} for tempApi, reuse File(), StateFlow.first for
    pause, Job tracking for coroutines, Dispatchers.Main for browse()
  - AppReducer: cap statusLog at 500, align filter, store upload percent
  - AppStore: Dispatchers.Main, remove dead collectAsState
  - DropZone: remove redundant color in style.copy
  - StatusReport/FileList: add key parameter to items()
  - ErrorHandler: null-safe error.message
  - ProgressSection: handle Double percent from VTModels change

  Kotlin idioms (#1-45):
  - Rename envFloat to envDouble, top-level const val for ColorTags
  - camelCase CacheEntry fields with @SerialName
  - isCleared as property, @Deprecated on finalize
  - Instant.now() instead of LocalDateTime, Float to Double
  - String.format simplification, @TempDir replacement
  - ApiKeyResolution data class, shared QuotaManager in CacheCommand
  - PARTIAL_SUCCESS exit code, explicit imports over star imports
  - OutputFormatter: PRETTY_JSON constant, regular function over
    extension on Any?, warn on unknown types, validate output format

  CLI fixes (2.5-2.8):
  - Wrap api in try/finally across all commands
  - Remove System.getenv("USER") fallback
  - Wire --no-recursive flag in ScanCommand
  - Fix QuotaCommand hardcoded "user" literal

  Build cleanup:
  - Remove 10 unused version properties from gradle.properties
  - Remove unused okhttp/kotlinx-datetime from libs.versions.toml
  - Remove redundant kotlin-stdlib dependencies

  Bug fixes found during review:
  - Fix UserInfo.kt missing ContentNegotiation plugin (NoTransformationFoundException)
  - Fix StatusReport duplicate key crash on repeated log messages
  - Fix CliCommands.kt callers after QuotaManager field rename

+++ desktop/src/main/kotlin/com/vtbatch/desktop/ui/components/ProgressSection.kt
@@ -40,7 +40,7 @@ private fun ProgressRow(
             modifier = Modifier.width(80.dp)
         )
         LinearProgressIndicator(
-            progress = { progress.percent },
+            progress = { progress.percent.toFloat() },
             modifier = Modifier.weight(1f).height(8.dp),
             color = color,
             trackColor = MaterialTheme.colorScheme.surfaceVariant,
